# tests/openapi.yaml
openapi: 3.0.3
info:
  title: Tiikii Festival API
  description: API para el sistema de festivales Tiikii con funcionalidades de tickets, chat, mesh network y más
  version: 1.0.0
  contact:
    name: Tiikii Development Team
    email: dev@tiikii.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api.tiikii.com
    description: Servidor de producción
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      summary: Registrar usuario
      description: Crea una nueva cuenta de usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, firstName, lastName]
              properties:
                username:
                  type: string
                  example: "testuser"
                email:
                  type: string
                  format: email
                  example: "test@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "password123"
                firstName:
                  type: string
                  example: "Test"
                lastName:
                  type: string
                  example: "User"
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/auth/login:
    post:
      summary: Iniciar sesión
      description: Autentica un usuario y devuelve un token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "test@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refreshToken:
                        type: string
                      expiresIn:
                        type: string
                        example: "7d"
                  message:
                    type: string
                    example: "Login successful"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/me:
    get:
      summary: Obtener perfil del usuario
      description: Devuelve la información del usuario autenticado
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: "User profile retrieved successfully"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          example: "user-123"
        username:
          type: string
          example: "testuser"
        email:
          type: string
          format: email
          example: "test@example.com"
        firstName:
          type: string
          example: "Test"
        lastName:
          type: string
          example: "User"
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
          example: true
        isVerified:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              example: "ValidationError"
            message:
              type: string
              example: "Invalid input data"
            code:
              type: string
              example: "VALIDATION_ERROR"
            field:
              type: string
              example: "email"
            details:
              type: array
              items:
                type: string
              example: ["Field 'email' is required"]
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "ValidationError"
              message: "Invalid input data"
              code: "VALIDATION_ERROR"
              details: ["Field 'email' is required"]
    UnauthorizedError:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "UnauthorizedError"
              message: "Authentication required"
              code: "UNAUTHORIZED"
...